requirements.txt
Flask
Pillow
pyzbar
qrcode[pil]